+++ modules/c++/math.poly/include/math/poly/TwoD.h
@@ -64,7 +64,7 @@ public:
 
     TwoD(size_t orderX, size_t orderY) : mCoef(orderX+1,OneD<_T>(orderY)) {}   
 
-    template<typename Vector_T> TwoD(int orderX, int orderY,
+    template<typename Vector_T> TwoD(size_t orderX, size_t orderY,
                                      const Vector_T& coeffs)
     {
         mCoef.resize(orderX+1,OneD<_T>(orderY));
